+++ lib/data/campaigns.ts
@@ -350,13 +350,25 @@ export async function getCampaignById(id: string) {
       title: result.campaign.title,
       description: result.campaign.description,
       guidelines: result.campaign.guidelines,
+      ipKitId: result.campaign.ipKitId,
       brand_name: result.brand?.name,
       status: result.campaign.status,
+      startDate: result.campaign.startDate,
+      endDate: result.campaign.endDate,
       deadline: result.campaign.endDate,
+      maxSubmissions: result.campaign.maxSubmissions,
+      rewardAmount: result.campaign.rewardAmount,
+      rewardCurrency: result.campaign.rewardCurrency,
+      briefDocument: result.campaign.briefDocument,
       created_at: result.campaign.createdAt,
       featured: result.campaign.featuredUntil ? new Date(result.campaign.featuredUntil) > new Date() : false,
       brand: result.brand ? {
         name: result.brand.name
+      } : null,
+      ipKit: result.ipKit ? {
+        id: result.ipKit.id,
+        name: result.ipKit.name,
+        description: result.ipKit.description,
       } : null,
       createdAt: result.campaign.createdAt,
       updatedAt: result.campaign.updatedAt,
